requirements.txt
Flask==3.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
PyJWT==2.8.0
